+++ src/algorithm/heapSort.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include <iostream>
+#include <algorithm>
 
 void heapify(int*, int, int);
 void heapSort(int*, int);
-void heapify(int*, int, int, int&);
-void heapSort(int*, int, int&);
+void heapify(int*, int, int, long long&);
+void heapSort(int*, int, long long&);
