+++ src/algorithm/mergeSort.h
@@ -1,8 +1,9 @@
 #pragma once
 
+#include <algorithm>
 #include <limits.h>
 
 void merge(int*, int, int, int);
 void mergeSort(int*, int);
-void merge(int*, int, int, int, int&);
-void mergeSort(int*, int, int&);
+void merge(int*, int, int, int, long long&);
+void mergeSort(int*, int, long long&);
