+++ src/algorithm/quickSort.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include <iostream>
+#include <algorithm>
 
 int partition(int*, int, int);
 void quickSort(int*, int);
-int partition(int*, int, int, int&);
-void quickSort(int*, int, int&);
+int partition(int*, int, int, long long&);
+void quickSort(int*, int, long long&);
